+++ modules/@apostrophecms/notification/ui/apos/components/AposNotification.vue
@@ -2,6 +2,7 @@
   <div
     role="alert"
     :class="classList"
+    data-apos-notification=""
   >
     <span class="apos-notification__indicator">
       <AposIndicator