+++ modules/@apostrophecms/rich-text-widget/ui/apos/components/AposRichTextWidgetEditor.vue
@@ -383,7 +383,6 @@ export default {
           this.emitWidgetUpdate();
         }
       } else {
-        console.log('focus change emit');
         apos.bus.$emit('close-context-menus');
       }
     },
@@ -488,7 +487,6 @@ export default {
       this.closeToolbar();
     },
     onBubbleHide() {
-      console.log('bubble hide emit');
       apos.bus.$emit('close-context-menus', 'richText');
     },
     generateKey() {
@@ -496,8 +494,9 @@ export default {
         Math.random().toString(36).substring(2, 15);
     },
     handleUIKeydown(e) {
-      if (!e.aposConsumedEscape && (e.key === 'Escape')) {
-        e.aposConsumedEscape = true;
+      if (e.key === 'Escape') {
+        // Don't confuse escape key handlers in other modal layers etc.
+        e.stopPropagation();
         this.doSuppressInsertMenu();
       } else {
         this.suppressInsertMenu = false;