+++ modules/@apostrophecms/rich-text-widget/ui/apos/components/AposTiptapLink.vue
@@ -55,6 +55,7 @@
 
 <script>
 import AposEditorMixin from 'Modules/@apostrophecms/modal/mixins/AposEditorMixin';
+import { useModalStore } from 'Modules/@apostrophecms/ui/stores/modal';
 
 export default {
   name: 'AposTiptapLink',
@@ -147,6 +148,7 @@ export default {
     }
   },
   async mounted() {
+    this.modalStore = useModalStore();
     await this.evaluateExternalConditions();
   },
   methods: {
@@ -219,8 +221,9 @@ export default {
       this.close();
     },
     keyboardHandler(e) {
-      if (!e.aposConsumedEscape && (e.key === 'Escape')) {
-        e.aposConsumedEscape = true;
+      if (e.key === 'Escape') {
+        // Don't confuse escape key handlers in other modal layers etc.
+        e.stopPropagation();
         this.close();
       }
       if (e.key === 'Enter') {
@@ -299,14 +302,14 @@ export default {
     },
     async openPopover() {
       this.hasLinkOnOpen = Boolean(this.attributes.href);
-      window.addEventListener('keydown', this.keyboardHandler);
+      this.modalStore.onKeyDown(this.$el, this.keyboardHandler);
       await this.populateFields();
       this.evaluateConditions();
       apos.adminBar.disableRefresh();
       this.$emit('open-popover');
     },
     closePopover() {
-      window.removeEventListener('keydown', this.keyboardHandler);
+      this.modalStore.offKeyDown(this.keyboardHandler);
       apos.adminBar.enableRefresh();
       this.$emit('close');
     }