+++ modules/@apostrophecms/ui/ui/apos/stores/modal.js
@@ -5,6 +5,9 @@ import { createId } from '@paralleldrive/cuid2';
 export const useModalStore = defineStore('modal', () => {
   const stack = ref([]);
   const activeId = ref(null);
+  let keyDownListeners = [];
+
+  window.addEventListener('keydown', handleKeyDown);
 
   const activeModal = computed(() => {
     return stack.value.find(modal => activeId.value === modal.id);
@@ -39,6 +42,33 @@ export const useModalStore = defineStore('modal', () => {
 
   function getDepth() {
     return stack.value.length;
+  }
+
+  // Listens for keystrokes globally, but delivers them only
+  // if "el" is currently in the topmost modal (if any)
+  function onKeyDown(el, fn) {
+    keyDownListeners.push({
+      el,
+      fn
+    });
+  }
+
+  // Reverse of onKeyDown. Note you call it with just fn
+  function offKeyDown(fn) {
+    if (arguments.length === 2) {
+      throw new Error('Call offKeyDown with just fn');
+    }
+    keyDownListeners = keyDownListeners.filter(({ fn: fnFound }) => fnFound !== fn);
+  }
+
+  function handleKeyDown(e) {
+    const top = stack.value.at(-1)?.modalEl;
+    for (const { el, fn } of keyDownListeners) {
+      if (top && (getDepthOf(el) < getDepthOf(top))) {
+        continue;
+      }
+      fn(e);
+    }
   }
 
   function getAt(index) {
@@ -146,8 +176,7 @@ export const useModalStore = defineStore('modal', () => {
   // If el2 is `document` and el1 is a modal, this
   // method returns true. For convenenience, if el1
   // or el2 is not a modal, it is treated as its DOM
-  // parent modal, or as `document`. If el1 has no
-  // parent modal this method always returns false.
+  // parent modal, or as `document`.
   //
   // If el1 is no longer connected to the DOM then it
   // is also considered to be "on top" e.g. not something
@@ -155,37 +184,44 @@ export const useModalStore = defineStore('modal', () => {
   // similar functionality. This is necessary because
   // sometimes Vue removes elements from the DOM before
   // we can examine their relationships.
+  //
+  // If el1 is part of a notification, it is always considered
+  // to be on top of el2. This prevents unwanted dismissals of
+  // intermediate modals beneath notifications.
   function onTopOf(el1, el2) {
+    if (el1.closest('[data-apos-notification]')) {
+      return true;
+    }
+    // Why is this here? Things can be stacked above context menus.
+    // But, I'm afraid to remove it -Tom
     if (el2.matches('[data-apos-menu]')) {
       return false;
     }
     if (!el1.isConnected) {
-    // If el1 is no longer in the DOM we can't make a proper determination,
-    // returning true prevents unwanted things like click-outside-element
-    // events from firing
+      // If el1 is no longer in the DOM we can't make a proper determination,
+      // returning true prevents unwanted things like click-outside-element
+      // events from firing
       return true;
     }
-    if (!el1.matches('[data-apos-modal]')) {
-      el1 = el1.closest('[data-apos-modal]') || document;
-    }
-    if (!el2.matches('[data-apos-modal]')) {
-      el2 = el2.closest('[data-apos-modal]') || document;
-    }
-    if (el1 === document) {
-      return false;
-    }
-    if (el2 === document) {
-      return true;
-    }
-    const index1 = stack.value.findIndex(modal => modal.modalEl === el1);
-    const index2 = stack.value.findIndex(modal => modal.modalEl === el2);
-    if (index1 === -1) {
-      throw new Error('apos.modal.onTopOf: el1 is not in the modal stack');
-    }
-    if (index2 === -1) {
-      throw new Error('apos.modal.onTopOf: el2 is not in the modal stack');
-    }
+    const index1 = getDepthOf(el1);
+    const index2 = getDepthOf(el2);
     return index1 > index2;
+  }
+
+  // Returns the depth of el in the modal stack, where higher numbers
+  // are higher in the stack ("on top"). Returns -1 if el is not
+  // in any modal. If el is not in the DOM -1 is always returned
+  function getDepthOf(el) {
+    if (!el.isConnected) {
+      return -1;
+    }
+    if (!el.matches('[data-apos-modal]')) {
+      el = el.closest('[data-apos-modal]') || document;
+    }
+    if (el === document) {
+      return -1;
+    }
+    return stack.value.findIndex(modal => modal.modalEl === el);
   }
 
   return {
@@ -205,6 +241,8 @@ export const useModalStore = defineStore('modal', () => {
     confirm,
     alert,
     report,
-    onTopOf
+    onTopOf,
+    onKeyDown,
+    offKeyDown
   };
 });